Cosmetic changes in RabbitmqCluster
