+++ src/controller_examples/rabbitmq_controller/exec/mod.rs
@@ -1,3 +1,4 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
+pub mod rabbitmqcluster;
 pub mod reconciler;
